further

+++ src/component/gas_producer_component.cc
@@ -108,10 +108,10 @@ std::unique_ptr<component_stub> gas_producer_component_stub::from_config(const c
     gas_producer_stub->gas_type = config_setting_get_int(gas_type_member);
 
     auto flow_rate_member = config_setting_get_member(config, "flow_rate");
-    gas_producer_stub->flow_rate = config_setting_get_float(flow_rate_member);
+    gas_producer_stub->flow_rate = (float)config_setting_get_float(flow_rate_member);
 
     auto max_pressure_member = config_setting_get_member(config, "max_pressure");
-    gas_producer_stub->max_pressure = config_setting_get_float(max_pressure_member);
+    gas_producer_stub->max_pressure = (float)config_setting_get_float(max_pressure_member);
 
     return std::move(gas_producer_stub);
 }
