+++ src/component/light_component.cc
@@ -89,7 +89,7 @@ std::unique_ptr<component_stub> light_component_stub::from_config(const config_s
     auto light_stub = std::make_unique<light_component_stub>();
 
     auto intensity_member = config_setting_get_member(config, "intensity");
-    light_stub->intensity = config_setting_get_float(intensity_member);
+    light_stub->intensity = (float)config_setting_get_float(intensity_member);
 
     return std::move(light_stub);
 }
