+++ src/component/power_provider_component.cc
@@ -95,7 +95,7 @@ std::unique_ptr<component_stub> power_provider_component_stub::from_config(const
     auto power_provider_stub = std::make_unique<power_provider_component_stub>();
 
     auto max_provided_member = config_setting_get_member(config, "max_provided");
-    power_provider_stub->max_provided = config_setting_get_float(max_provided_member);
+    power_provider_stub->max_provided = (float)config_setting_get_float(max_provided_member);
 
     return std::move(power_provider_stub);
 }
