+++ src/component/sensor_comparator_component.cc
@@ -89,7 +89,7 @@ std::unique_ptr<component_stub> sensor_comparator_component_stub::from_config(co
     auto sensor_comparator_stub = std::make_unique<sensor_comparator_component_stub>();
 
     auto compare_epsilon_member = config_setting_get_member(config, "compare_epsilon");
-    sensor_comparator_stub->compare_epsilon = config_setting_get_float(compare_epsilon_member);
+    sensor_comparator_stub->compare_epsilon = (float)config_setting_get_float(compare_epsilon_member);
 
     return std::move(sensor_comparator_stub);
 }
